requirements.txt
aiohttp==3.5.4
alembic==1.3.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
discord.py==1.2.5
enum34==1.1.6
exif==0.8.1
gphotos-sync==2.10.3
idna==2.8
Mako==1.1.0
MarkupSafe==1.1.1
multidict==4.7.2
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
Pillow==6.2.1
pytesseract==0.3.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
pywin32==227
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.13.0
SQLAlchemy==1.3.12
urllib3==1.25.7
websockets==6.0
yarl==1.4.2
